refactor(wire): common
