+++ app/common/framework.go
@@ -1,0 +1,8 @@
+package common
+
+import "github.com/google/wire"
+
+var Framework = wire.NewSet(
+	wire.Struct(new(GlobalInitializer), "*"),
+	wire.Struct(new(Runner), "*"),
+)
