+++ app/common/ingest.go
@@ -1,0 +1,10 @@
+package common
+
+import (
+	"github.com/google/wire"
+)
+
+var Ingest = wire.NewSet(
+	NewKafkaIngestCollector,
+	NewIngestCollector,
+)
