+++ app/common/kafka.go
@@ -5,6 +5,7 @@ import (
 	"log/slog"
 
 	"github.com/confluentinc/confluent-kafka-go/v2/kafka"
+	"github.com/google/wire"
 	"go.opentelemetry.io/otel/metric"
 
 	"github.com/openmeterio/openmeter/app/config"
@@ -16,12 +17,43 @@ import (
 	kafkametrics "github.com/openmeterio/openmeter/pkg/kafka/metrics"
 )
 
+var KafkaConfig = wire.NewSet(
+	// TODO: refactor to move out Kafka config from ingest and consolidate
+	wire.FieldsOf(new(config.KafkaIngestConfiguration), "KafkaConfiguration"),
+	wire.FieldsOf(new(config.Configuration), "Ingest"),
+	wire.FieldsOf(new(config.IngestConfiguration), "Kafka"),
+)
+
+var Kafka = wire.NewSet(
+	NewKafkaProducer,
+	NewKafkaMetrics,
+
+	KafkaTopic,
+)
+
+var KafkaTopic = wire.NewSet(
+	wire.FieldsOf(new(config.KafkaIngestConfiguration), "TopicProvisionerConfig"),
+
+	NewKafkaAdminClient,
+
+	NewKafkaTopicProvisionerConfig,
+	NewKafkaTopicProvisioner,
+)
+
+var KafkaNamespaceResolver = wire.NewSet(
+	NewNamespacedTopicResolver,
+	wire.Bind(new(topicresolver.Resolver), new(*topicresolver.NamespacedTopicResolver)),
+
+	NewKafkaNamespaceHandler,
+	NewNamespaceHandlers,
+)
+
 // TODO: use ingest config directly?
 // TODO: use kafka config directly?
 // TODO: add closer function?
-func NewKafkaProducer(conf config.Configuration, logger *slog.Logger) (*kafka.Producer, error) {
+func NewKafkaProducer(conf config.KafkaIngestConfiguration, logger *slog.Logger) (*kafka.Producer, error) {
 	// Initialize Kafka Admin Client
-	kafkaConfig := conf.Ingest.Kafka.CreateKafkaConfig()
+	kafkaConfig := conf.CreateKafkaConfig()
 
 	// Initialize Kafka Producer
 	producer, err := kafka.NewProducer(&kafkaConfig)
@@ -111,8 +143,8 @@ func NewKafkaNamespaceHandler(
 	}, nil
 }
 
-func NewNamespacedTopicResolver(config config.Configuration) (*topicresolver.NamespacedTopicResolver, error) {
-	topicResolver, err := topicresolver.NewNamespacedTopicResolver(config.Ingest.Kafka.EventsTopicTemplate)
+func NewNamespacedTopicResolver(config config.KafkaIngestConfiguration) (*topicresolver.NamespacedTopicResolver, error) {
+	topicResolver, err := topicresolver.NewNamespacedTopicResolver(config.EventsTopicTemplate)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create topic name resolver: %w", err)
 	}
