+++ cmd/balance-worker/wire_gen.go
@@ -146,8 +146,8 @@ func initializeApplication(ctx context.Context, conf config.Configuration) (Appl
 		return Application{}, nil, err
 	}
 	v3 := conf.Meters
-	inMemoryRepository := common.NewMeterRepository(v3)
-	connector, err := common.NewStreamingConnector(ctx, aggregationConfiguration, v2, inMemoryRepository, logger)
+	repository := common.NewInMemoryRepository(v3)
+	connector, err := common.NewStreamingConnector(ctx, aggregationConfiguration, v2, repository, logger)
 	if err != nil {
 		cleanup6()
 		cleanup5()
@@ -161,13 +161,13 @@ func initializeApplication(ctx context.Context, conf config.Configuration) (Appl
 		DatabaseClient:     client,
 		StreamingConnector: connector,
 		Logger:             logger,
-		MeterRepository:    inMemoryRepository,
+		MeterRepository:    repository,
 		Publisher:          eventbusPublisher,
 	}
 	entitlement := registrybuilder.GetEntitlementRegistry(entitlementOptions)
-	entitlementRepo := common.NewEntitlementRepo(client)
+	balanceWorkerEntitlementRepo := common.NewBalanceWorkerEntitlementRepo(client)
 	subjectResolver := common.SubjectResolver()
-	workerOptions := common.NewBalanceWorkerOptions(eventsConfiguration, options, eventbusPublisher, entitlement, entitlementRepo, subjectResolver, logger)
+	workerOptions := common.NewBalanceWorkerOptions(eventsConfiguration, options, eventbusPublisher, entitlement, balanceWorkerEntitlementRepo, subjectResolver, logger)
 	worker, err := common.NewBalanceWorker(workerOptions)
 	if err != nil {
 		cleanup6()
