+++ cmd/notification-service/wire.go
@@ -43,17 +43,19 @@ type Application struct {
 func initializeApplication(ctx context.Context, conf config.Configuration) (Application, func(), error) {
 	wire.Build(
 		metadata,
-		common.Config,
 		common.ClickHouse,
 		common.Database,
+		common.Event,
 		common.Feature,
 		common.Framework,
+		common.KafkaConfig,
 		common.KafkaTopic,
+		common.MeterInMemory,
 		common.NewDefaultTextMapPropagator,
 		common.Notification,
 		common.NotificationServiceProvisionTopics,
+		common.Streaming,
 		common.Svix,
-		common.OpenMeter,
 		common.Telemetry,
 		common.Watermill,
 		wire.Struct(new(Application), "*"),
