+++ cmd/notification-service/wire_gen.go
@@ -123,8 +123,8 @@ func initializeApplication(ctx context.Context, conf config.Configuration) (Appl
 		return Application{}, nil, err
 	}
 	v2 := conf.Meters
-	inMemoryRepository := common.NewMeterRepository(v2)
-	featureConnector := common.NewFeatureConnector(logger, client, inMemoryRepository)
+	repository := common.NewInMemoryRepository(v2)
+	featureConnector := common.NewFeatureConnector(logger, client, repository)
 	v3 := conf.Svix
 	service, err := common.NewNotificationService(logger, client, notificationConfiguration, v3, featureConnector)
 	if err != nil {
@@ -148,7 +148,7 @@ func initializeApplication(ctx context.Context, conf config.Configuration) (Appl
 		cleanup()
 		return Application{}, nil, err
 	}
-	connector, err := common.NewStreamingConnector(ctx, aggregationConfiguration, v4, inMemoryRepository, logger)
+	connector, err := common.NewStreamingConnector(ctx, aggregationConfiguration, v4, repository, logger)
 	if err != nil {
 		cleanup6()
 		cleanup5()
@@ -172,7 +172,7 @@ func initializeApplication(ctx context.Context, conf config.Configuration) (Appl
 		MessagePublisher:   publisher,
 		Meter:              meter,
 		Metadata:           commonMetadata,
-		MeterRepository:    inMemoryRepository,
+		MeterRepository:    repository,
 		Notification:       service,
 		StreamingConnector: connector,
 		TelemetryServer:    v5,
